refactor: fix the tests having wrong directory
